layout/index.html
This template gets rendered when we hit the root URL of a Hugo site.

We define the content of layout/index.html enclosed within a define "main" block.
Since Hugo always renders layout/_default/baseof.html, the content of
this definition of "main" block will be injected into the baseof.html's "main" block

That way we get the header, footer (and all other content of
baseof.html) for free.